Move the generate style variables to StyleDictionary
