+++ src/index.ts
@@ -1,22 +1,17 @@
-import styleDictionary from '@/styleDictionary/styleDictionary.ts';
+import { buildStyleDictionary } from '@/styleDictionary/StyleDictionary.ts';
 import { Arguments } from 'yargs-parser';
 import { ParseConfig } from '@/types/designTokens';
 import { figmaApiConnection } from '@/api';
-import { createCssTokenFiles, createJsonTokenFiles, EMOJIS, log } from '@/functions';
+import { createJsonTokenFiles, EMOJIS, log } from '@/functions';
 
 const designTokens = (args: Arguments, config: ParseConfig) => {
-  const command = args._[0];
   const { settings, configFile } = config;
 
   figmaApiConnection(settings).then(async generatedTokens => {
     log('Generating design tokens...', EMOJIS.workingInProgress);
 
     await createJsonTokenFiles(generatedTokens, settings);
-    await createCssTokenFiles(generatedTokens, settings);
-
-    if (command === 'platforms') {
-      styleDictionary(configFile);
-    }
+    buildStyleDictionary(configFile);
   });
 };
