+++ src/types/designTokens/color.ts
@@ -8,9 +8,9 @@ export interface ColorCollection extends TokenCollection {
 
 export interface Color {
   name: string;
-  rgbColor: RgbColor;
-  hexColor: string;
-  hslColor: HslColor;
+  value: string;
+  valueRgb: RgbColor;
+  valueHsl: HslColor;
 }
 
 export interface HslColor {
