+++ src/types/designTokens/designTokens.ts
@@ -4,6 +4,6 @@ import { TokenCollection } from '@/types/designTokens/tokens.ts';
 export interface DesignTokensGenerator {
   name: string;
   tokens: TokenCollection;
+
   writeTokens(createFile: CreateFile, outputDir: string, name?: string): Promise<void>[];
-  writeCssVariables(createFile: CreateFile, outputDir: string, name?: string): Promise<void>[];
 }
